# fix: Replace flaky performance test with robust algorithmic complexity test